# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "One Medy API Documentation",
        "description": "One Medy APi Service to be consumed by client Application",
        "contact": {
            "email": "info@walexbiznig.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/bank-detail/add": {
            "post": {
                "tags": [
                    "Bank Detail"
                ],
                "summary": "user adds his bank details",
                "operationId": "be8e5f66510be647fd998add5d3ce9db",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBankDetailRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/bank-detail/byid/": {
            "get": {
                "tags": [
                    "Bank Detail"
                ],
                "summary": "Get bank detail by user id",
                "description": "Returns the bank detail of the specified user",
                "operationId": "getById",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Size"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/v1/doctorspecialization/create": {
            "post": {
                "tags": [
                    "Doctor"
                ],
                "summary": "doctors add thier specializations",
                "operationId": "a5549c50e0a788539f22d31e87ad4303",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDoctorSpecializationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/doctorsearch": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "summary": "to get doctor detail",
                "operationId": "2e0227de60aca5fa448dd22c1ffebc21",
                "parameters": [
                    {
                        "name": "specialization_id",
                        "in": "query",
                        "description": "For SPecialization"
                    },
                    {
                        "name": "availability",
                        "in": "query",
                        "description": "For SPecialization"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/doctor/specialization/{id}": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "summary": "To get the doctor's specializtation",
                "operationId": "601bd28072a64cbbcbd742029250f8c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "doctor id",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/api/v1/drug/prescription/byappointment/": {
            "get": {
                "tags": [
                    "Drug Prescription"
                ],
                "summary": "Get prescription by appointment id",
                "description": "Returns the prescription the specified user",
                "operationId": "getByAppointmentId",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Size"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateBankDetailRequest": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "account_number": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "bank": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "sort_code": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/CreateBankDetailRequest"
                    }
                ]
            },
...
            "CreateDoctorSpecializationRequest": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "specialization_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "doctor_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/CreateDoctorSpecializationRequest"
                    }
                ]
            },
...
        },
...
    },
...
    "security": [
        []
    ]
}